backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
requires-python = ">=3.10,<3.11"
dependencies = [
    "fastapi[standard]<1.0.0,>=0.114.2",
    "python-multipart<1.0.0,>=0.0.7",
    "pydantic>2.0",
    "pydantic-settings<3.0.0,>=2.2.1",
    "pymongo==4.11",
    "odmantic>=1.0.2",
    "pyorthanc>=1.19.1",
    "httpx>=0.27.0",
    "pydicom>=2.4.4",
    "numpy>=2.2.3",
    "matplotlib>=3.10.1",
    "requests>=2.32.3",
    "simpleitk>=2.3.1",
    "vtk>=9.4.1",
    "scipy>=1.15.2",
    "pynrrd>=1.1.3",
    "opencv-python>=4.11.0.86",
]

[tool.uv]
dev-dependencies = [
    "pytest<8.0.0,>=7.4.3",
    "pytest-asyncio>=0.23.8",
    "mypy<2.0.0,>=1.8.0",
    "ruff<1.0.0,>=0.2.2",
    "isort>=6.0.1",
    "pre-commit<4.0.0,>=3.6.2",
    "types-passlib<2.0.0.0,>=1.7.7.20240106",
    "coverage<8.0.0,>=7.4.3",
    "ipykernel>=6.29.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.mypy]
strict = true
exclude = ["venv", ".venv"]

[tool.ruff]
target-version = "py310"
exclude = []

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
